C++
If If a and b are coprime, then c=gcd(a,b)=1. Then the following equation holds: a+b+1=c. We know that for any integer a, gcd(a−1,a)=1 holds. if a is an even number, then gcd(a−1,a+1) holds. if a is an odd number, then gcd(a−2,a+2) holds.